BACKEND/BACKEND/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-multipart
pdfplumber
pypdf
deepseek
python-dotenv
python-jose
sqlalchemy
passlib
pymysql
PyMuPDF
